bug fix for multiple set indices
I ran into an issue when attempting to index multiple sets using an array of integers for the second argument. On line 57, "set_in < 0" returns an array of logical values, when || can only accept a single logical. I'm not sure if my solution will work for the intended implementation, but it worked perfectly for me.